check for integer oveflow

+++ coders/wpg.c
@@ -684,7 +684,7 @@ static int UnpackWPG2Raster(Image *image,int bpp)
                   InsertByte6(SampleBuffer[bbuf]);
             }
           else {      /* NRP */
-            for(i=0; i< SampleSize*(RunCount+1);i++)
+            for(i=0; i < (ssize_t) (SampleSize*(RunCount+1)); i++)
               {
                 bbuf=ReadBlobByte(image);
                 InsertByte6(bbuf);
